+++ web/__init__.py
@@ -1,0 +1,5 @@
+"""Web package."""
+
+from fastapi.templating import Jinja2Templates
+
+templates = Jinja2Templates(directory="./web/templates")
