+++ web/public/favicon/android-chrome-192x192.png
