+++ web/templates/base.html.jinja
@@ -9,8 +9,8 @@
 	<meta name="viewport" content="width=device-width, initial-scale=1">
 	<meta name="color-scheme" content="light dark">
 	<link href="https://cdn.jsdelivr.net/npm/bootstrap-dark-5@1.1.3/dist/css/bootstrap-dark.min.css" rel="stylesheet">
-	<link href="/web/style.css" rel="stylesheet">
-	<link rel="icon" type="image/x-icon" href="/web/favicon/favicon.ico">
+	<link href="/public/style.css" rel="stylesheet">
+	<link rel="icon" type="image/x-icon" href="/public/favicon/favicon.ico">
 
 	<!-- Meta descriptions -->
 	<meta property="og:type" content="website">
@@ -40,8 +40,8 @@
 			<div class="col text-center text-muted">
 				<hr>
 				<strong>Inconnu</strong> is copyright &copy;2023 by tiltowait. All rights reserved.<br>
-				View the source code <a href="https://github.com/tiltowait/inconnu" title="Inconnu source code">here</a>.<br>
-				Want to support me? Join the <a href="https://www.patreon.com/tiltowait" title="Patreon">Patreon</a>!
+				<a href="https://github.com/tiltowait/inconnu" title="Inconnu source code">View the source code</a> on GitHub.<br>
+				Want to support <strong>Inconnu's</strong> development? <a href="https://www.patreon.com/tiltowait" title="Patreon">Join the Patreon!</a>
 			</div>
 			{% endblock %}
 		</div>
